+++ acount-rest-service/handler/category.go
@@ -3,7 +3,6 @@ package handler
 import (
 	"database/sql"
 	"encoding/json"
-	"log"
 	"net/http"
 	"strconv"
 	"strings"
@@ -14,90 +13,10 @@ import (
 	"github.com/paypay3/kakeibo-app-api/acount-rest-service/domain/model"
 
 	"github.com/garyburd/redigo/redis"
-
-	"github.com/paypay3/kakeibo-app-api/acount-rest-service/domain/repository"
 )
-
-type DBHandler struct {
-	DBRepo repository.DBRepository
-}
 
 type DeleteCustomCategoryMsg struct {
 	Message string `json:"message"`
-}
-
-type HTTPError struct {
-	Status       int   `json:"status"`
-	ErrorMessage error `json:"error"`
-}
-
-type AuthenticationErrorMsg struct {
-	Message string `json:"message"`
-}
-
-type ValidationErrorMsg struct {
-	Message string `json:"message"`
-}
-type ConflictErrorMsg struct {
-	Message string `json:"message"`
-}
-
-type InternalServerErrorMsg struct {
-	Message string `json:"message"`
-}
-
-func NewDBHandler(DBRepo repository.DBRepository) *DBHandler {
-	DBHandler := DBHandler{DBRepo: DBRepo}
-	return &DBHandler
-}
-
-func NewHTTPError(status int, err error) error {
-	switch status {
-	case http.StatusBadRequest:
-		return &HTTPError{
-			Status:       status,
-			ErrorMessage: err.(*ValidationErrorMsg),
-		}
-	case http.StatusConflict:
-		return &HTTPError{
-			Status:       status,
-			ErrorMessage: err.(*ConflictErrorMsg),
-		}
-	case http.StatusUnauthorized:
-		return &HTTPError{
-			Status:       status,
-			ErrorMessage: &AuthenticationErrorMsg{"このページを表示するにはログインが必要です"},
-		}
-	default:
-		return &HTTPError{
-			Status:       status,
-			ErrorMessage: &InternalServerErrorMsg{"500 Internal Server Error"},
-		}
-	}
-}
-
-func (e *HTTPError) Error() string {
-	b, err := json.Marshal(e)
-	if err != nil {
-		log.Println(err)
-	}
-	return string(b)
-}
-
-func (e *ValidationErrorMsg) Error() string {
-	return e.Message
-}
-
-func (e *ConflictErrorMsg) Error() string {
-	return e.Message
-}
-
-func (e *AuthenticationErrorMsg) Error() string {
-	return e.Message
-}
-
-func (e *InternalServerErrorMsg) Error() string {
-	return e.Message
 }
 
 func validateCustomCategory(r *http.Request, customCategory *model.CustomCategory) error {
@@ -122,54 +41,6 @@ func validateCustomCategory(r *http.Request, customCategory *model.CustomCategor
 	}
 
 	return nil
-}
-
-func responseByJSON(w http.ResponseWriter, r *http.Request, data interface{}, err error) {
-	if err != nil {
-		httpError, ok := err.(*HTTPError)
-		if !ok {
-			w.WriteHeader(http.StatusInternalServerError)
-			return
-		}
-		w.Header().Set("Content-Type", "application/json; charset=UTF-8")
-		w.WriteHeader(httpError.Status)
-		if err := json.NewEncoder(w).Encode(httpError); err != nil {
-			w.WriteHeader(http.StatusInternalServerError)
-		}
-
-		return
-	}
-
-	if r.Method == http.MethodPost {
-		w.Header().Set("Content-Type", "application/json; charset=UTF-8")
-		w.WriteHeader(http.StatusCreated)
-		if err := json.NewEncoder(w).Encode(data); err != nil {
-			w.WriteHeader(http.StatusInternalServerError)
-			return
-		}
-
-		return
-	}
-
-	w.Header().Set("Content-Type", "application/json; charset=UTF-8")
-	w.WriteHeader(http.StatusOK)
-	if err := json.NewEncoder(w).Encode(data); err != nil {
-		w.WriteHeader(http.StatusInternalServerError)
-		return
-	}
-}
-
-func verifySessionID(h *DBHandler, w http.ResponseWriter, r *http.Request) (string, error) {
-	cookie, err := r.Cookie("session_id")
-	if err != nil {
-		return "", err
-	}
-	sessionID := cookie.Value
-	userID, err := h.DBRepo.GetUserID(sessionID)
-	if err != nil {
-		return "", err
-	}
-	return userID, nil
 }
 
 func (h *DBHandler) GetCategoriesList(w http.ResponseWriter, r *http.Request) {
